Split up services and endpoints
Curriculum service and -endpoint is now split up.
