+++ src/main/java/org/fsg1/fmms/backend/app/AppBinder.java
@@ -2,8 +2,9 @@ package org.fsg1.fmms.backend.app;
 
 import org.fsg1.fmms.backend.database.Connection;
 import org.fsg1.fmms.backend.services.CurriculaService;
-import org.fsg1.fmms.backend.services.CurriculumService;
+import org.fsg1.fmms.backend.services.ModulesService;
 import org.fsg1.fmms.backend.services.QualificationsService;
+import org.fsg1.fmms.backend.services.SemestersService;
 import org.glassfish.jersey.internal.inject.AbstractBinder;
 
 import javax.inject.Singleton;
@@ -19,7 +20,8 @@ public final class AppBinder extends AbstractBinder {
      */
     @Override
     protected void configure() {
-        bind(CurriculumService.class).to(CurriculumService.class).in(Singleton.class);
+        bind(SemestersService.class).to(SemestersService.class).in(Singleton.class);
+        bind(ModulesService.class).to(ModulesService.class).in(Singleton.class);
         bind(CurriculaService.class).to(CurriculaService.class).in(Singleton.class);
         bind(Connection.class).to(Connection.class).in(Singleton.class);
         bind(QualificationsService.class).to(QualificationsService.class).in(Singleton.class);
